Removing Erase logic from type inference
