+++ sol-core.cabal
@@ -64,6 +64,7 @@ library
       Solcore.Frontend.Syntax.Name
       Solcore.Frontend.Syntax.Stmt
       Solcore.Frontend.Syntax.SyntaxTree
+      Solcore.Frontend.TypeInference.Erase
       Solcore.Frontend.TypeInference.Id
       Solcore.Frontend.TypeInference.NameSupply
       Solcore.Frontend.TypeInference.SccAnalysis
